+++ DarkCodex/Patches/Patch_SaveExtension.cs
@@ -15,7 +15,7 @@ using CodexLib;
 
 namespace DarkCodex
 {
-    [PatchInfo(Severity.Harmony, "Save Metadata", "will add extra metadata of enabled patches and warn if you try to load a game with missing patches", false)] // TODO: rework
+    [PatchInfo(Severity.Harmony, "Save Metadata", "will add extra metadata of enabled patches and warn if you try to load a game with missing patches", false)]
     [HarmonyPatch]
     public class Patch_SaveExtension
     {
@@ -27,14 +27,17 @@ namespace DarkCodex
         {
             try
             {
+                Main.Print($"DEBUG OnSave 1");
+
                 if (!Settings.State.saveMetadata || Main.appliedPatches == null || Main.appliedPatches.Count == 0)
                     return;
+                Main.Print($"DEBUG OnSave 2 {Main.appliedPatches.Join()}");
 
                 string json = Helper.Serialize(Main.appliedPatches, indent: false, type: false);
                 if (json == null || json == "")
                     return;
 
-                saveInfo.Saver.SaveBytes(SaveKey, Encoding.Default.GetBytes(json));
+                saveInfo.Saver.SaveBytes(SaveKey, Encoding.Unicode.GetBytes(json));
                 Main.Print("Saved patch metadata");
             }
             catch (Exception e)
@@ -43,70 +46,60 @@ namespace DarkCodex
             }
         }
 
-        [HarmonyPatch(typeof(ThreadedGameLoader), nameof(ThreadedGameLoader.DoLoad))]
+        //[HarmonyPatch(typeof(ThreadedGameLoader), nameof(ThreadedGameLoader.DoLoad))]
+        [HarmonyPatch(typeof(Game), nameof(Game.LoadGame))]
         [HarmonyPrefix]
-        public static void OnLoad(ThreadedGameLoader __instance)
-        {
-            // no use
-        }
-
-        [HarmonyPatch(typeof(SaveSlot), nameof(SaveSlot.OnButtonSaveLoad))]
-        [HarmonyPriority(Priority.HigherThanNormal)]
-        [HarmonyPrefix]
-        public static bool BeforeLoad(SaveSlot __instance)
+        public static void OnLoad2(SaveInfo saveInfo, Game __instance)
         {
             try
             {
                 if (!Settings.State.saveMetadata)
-                    return true;
-
-                if (__instance.SlotType != SaveSlotType.Load)
-                    return true;
+                    return;
 
                 if (Main.restart)
                 {
-                    Helper.ShowMessageBox("Patch settings were changed recently. Restart game to apply new patches.", yesLabel: "Ignore this time", noLabel: "I understand",
+                    Helper.ShowMessageBox("Patch settings were changed recently. You must restart game now!", yesLabel: "Ignore this time", noLabel: "I understand",
                         onYes: () =>
                         {
                             Main.restart = false;
                         });
-                    return false;
                 }
 
-                var bytes = __instance.SaveInfo?.Saver?.ReadBytes(SaveKey);
+                var bytes = saveInfo?.Saver?.ReadBytes(SaveKey);
                 if (bytes == null)
-                    return true;
+                    return;
 
-                var json = Encoding.Default.GetString(bytes);
+                var json = Encoding.Unicode.GetString(bytes);
                 if (json == null || json == "")
-                    return true;
+                    return;
 
                 var saveData = Helper.Deserialize<List<string>>(value: json);
                 if (saveData == null)
-                    return true;
+                    return;
 
                 foreach (string patch in Main.appliedPatches)
                     saveData.Remove(patch);
 
                 if (saveData.Count > 0)
                 {
-                    Helper.ShowMessageBox("Critical patch missing! Either turn off 'Save Metadata' or press 'Enable' to close the game and enable: " + saveData.Join(), yesLabel: "Quit Game & Enable",
+                    Helper.ShowMessageBox("Critical patch missing! Either turn off 'Save Metadata' or press 'Enable patches' to enable: " + saveData.Join(), yesLabel: "Enable patches", noLabel: "Ignore this time",
                         onYes: () =>
                         {
                             foreach (var info in saveData)
                                 Main.patchInfos.SetEnable(true, info, force: true);
                             Main.OnSaveGUI(null);
+                            Helper.ShowMessageBox("Patches enabled, you must restart game now!");
                             //SystemUtil.ApplicationQuit();
-                            SystemUtil.ApplicationRestart();
+                            //SystemUtil.ApplicationRestart();
                         });
-                    return false;
+                    return;
                 }
-                return true;
+                return;
             }
             catch (Exception e)
             {
                 Main.PrintException(e);
-                return true;
+                return;
             }
         }
     }